requirements.txt
icalendar
requests
Pillow~=10.0.1